pyproject.toml
[tool.poetry]
name = "volleyball-uploader"
version = "0.0.0"
description = ""
authors = ["robkulesa <robjkulesa@gmail.com>"]
readme = "README.md"

packages = [
  {include = "src"}
]

[tool.poetry.dependencies]
python = "^3.12"
psycopg2-binary = "^2.9.9"
google-api-python-client = "^2.136.0"
google-auth-oauthlib = "^1.2.0"
oauth2client = "^4.1.3"
pandas = "^2.2.2"


[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"

[tool.black]
line-length = 180
target-version = ['py312']
preview = true

[tool.poetry.scripts]
volleyball-uploader = "src.main:main"

[tool.poetry-dynamic-versioning]
enable = true
